Apply new implementation of "list/task lists"
In the previous branch 65-api-structure "list/task lists" was
handled by using arrays filled with objects. That resulted in
bloated code when trying to apply changes to a list/task object
as each operation had to loop thought the array that contained
the list/task object

With the new implementation of list/task being object based,
it becomes much easier to apply changes to a list/task object.
As you only have to select the key of the object you are trying
to change